Organize deformation APIs into dedicated packages
