+++ test/nn/functional/geometry/test_point_deform.py
@@ -14,7 +14,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Tests for dense displacement and compact Shepard point morphing."""
+"""Tests for dense displacement and compact Shepard point deformation."""
 
 import inspect
 from typing import Literal, get_type_hints
@@ -26,7 +26,7 @@ import physicsnemo.nn.functional as functional
 from physicsnemo.core.function_spec import FunctionSpec
 from physicsnemo.nn.functional import displace_points, morph_points
 from physicsnemo.nn.functional.geometry import DisplacePoints, MorphPoints
-from physicsnemo.nn.functional.geometry.morphing import morphing as morphing_module
+from physicsnemo.nn.functional.geometry.deform import deform as deform_module
 from test.conftest import requires_module
 from test.nn.functional._parity_utils import clone_case
 
@@ -73,6 +73,12 @@ def _differentiable_case_tensors(args, kwargs):
 def test_public_exports_and_function_specs():
     assert displace_points.__name__ == "displace_points"
     assert morph_points.__name__ == "morph_points"
+    assert displace_points.__module__ == (
+        "physicsnemo.nn.functional.geometry.deform.deform"
+    )
+    assert morph_points.__module__ == (
+        "physicsnemo.nn.functional.geometry.deform.deform"
+    )
     assert issubclass(DisplacePoints, FunctionSpec)
     assert issubclass(MorphPoints, FunctionSpec)
     assert not hasattr(functional, "DisplacePoints")
@@ -743,7 +749,7 @@ def test_torch_chunk_checkpoint_bounds_retained_autograd_storage(monkeypatch):
     import gc
     import weakref
 
-    from physicsnemo.nn.functional.geometry.morphing import _torch_impl
+    from physicsnemo.nn.functional.geometry.deform import _torch_impl
 
     # Force several small query/control blocks without making the test expensive.
     monkeypatch.setattr(_torch_impl, "_PAIRWISE_TEMPORARY_BYTE_BUDGET", 32 * 1024)
@@ -813,7 +819,7 @@ def test_torch_control_chunked_gradients_match_unchunked_on_ties(monkeypatch):
     gradient-invariant.
     """
 
-    from physicsnemo.nn.functional.geometry.morphing import _torch_impl
+    from physicsnemo.nn.functional.geometry.deform import _torch_impl
 
     dtype = torch.float64
     points = torch.tensor([[0.0, 0.0], [0.1, 0.2]], dtype=dtype)
@@ -849,7 +855,7 @@ def test_compiled_torch_training_checkpoints_pairwise_activations():
     import gc
     import weakref
 
-    from physicsnemo.nn.functional.geometry.morphing._torch_impl import (
+    from physicsnemo.nn.functional.geometry.deform._torch_impl import (
         compact_shepard_field_torch,
     )
 
@@ -985,8 +991,8 @@ def test_default_dispatch_selects_device_backend(device, monkeypatch):
 
     # Patch the names resolved by the registered methods, rather than the source
     # implementation modules, to assert which custom dispatch branch ran.
-    monkeypatch.setattr(morphing_module, "morph_points_torch", torch_spy)
-    monkeypatch.setattr(morphing_module, "morph_points_warp", warp_spy)
+    monkeypatch.setattr(deform_module, "morph_points_torch", torch_spy)
+    monkeypatch.setattr(deform_module, "morph_points_warp", warp_spy)
 
     warp_impl = MorphPoints._get_impls()["warp"]
     expected = "warp" if device.type == "cuda" and warp_impl.available else "torch"
@@ -1016,7 +1022,7 @@ def test_default_dispatch_selects_device_backend(device, monkeypatch):
 
 @requires_module("warp")
 def test_warp_custom_ops_opcheck():
-    from physicsnemo.nn.functional.geometry.morphing._warp_impl import (
+    from physicsnemo.nn.functional.geometry.deform._warp_impl import (
         compact_shepard_field_warp_impl,
     )
 
@@ -1462,7 +1468,7 @@ def test_public_api_fake_tensor_propagation_with_tensor_options(
 def test_warp_raw_custom_op_fake_strides_match_noncontiguous_real_outputs(device):
     from torch._subclasses.fake_tensor import FakeTensorMode
 
-    from physicsnemo.nn.functional.geometry.morphing._warp_impl import (
+    from physicsnemo.nn.functional.geometry.deform._warp_impl import (
         compact_shepard_field_warp_impl,
     )
 
@@ -1491,7 +1497,7 @@ def test_warp_raw_custom_op_fake_strides_match_noncontiguous_real_outputs(device
 
 @requires_module("warp")
 def test_warp_no_grad_uses_forward_only_field(monkeypatch, device):
-    from physicsnemo.nn.functional.geometry.morphing._warp_impl import op as warp_op
+    from physicsnemo.nn.functional.geometry.deform._warp_impl import op as warp_op
 
     device = torch.device(device)
     points = torch.tensor(
@@ -1518,7 +1524,7 @@ def test_warp_no_grad_uses_forward_only_field(monkeypatch, device):
 def test_warp_zero_controls_bypass_launches_and_keep_zero_gradients(
     monkeypatch, device
 ):
-    from physicsnemo.nn.functional.geometry.morphing._warp_impl import op as warp_op
+    from physicsnemo.nn.functional.geometry.deform._warp_impl import op as warp_op
 
     device = torch.device(device)
     points = torch.randn((1, 4, 3), device=device, requires_grad=True)
@@ -1550,7 +1556,7 @@ def test_warp_zero_controls_bypass_launches_and_keep_zero_gradients(
 
 @requires_module("warp")
 def test_warp_control_displacement_only_pullback_returns_requested_gradient(device):
-    from physicsnemo.nn.functional.geometry.morphing._warp_impl import op as warp_op
+    from physicsnemo.nn.functional.geometry.deform._warp_impl import op as warp_op
 
     device = torch.device(device)
     points = torch.rand((1, 5, 3), device=device)
